Add Send Order

+++ src/app/shared/components/header/header.component.ts
@@ -27,12 +27,17 @@ export class HeaderComponent implements OnInit {
   }
 
   onLogout(): void{
-    let token = localStorage.getItem('token')
-    this.authService.logout(token+'').subscribe((res) => {
-      if (res) {
-        this.authService.deleteToken();
-        this.cookieService.delete('token_access');
-        this.isLogged = this.cookieService.check('token_access');
+    this.authService.updateToken().subscribe((r) => {
+      if (r) {
+        this.authService.saveToken(r.token, r.user)
+        this.cookieService.set('token_access', r.token, 1, '/');
+        let token = localStorage.getItem('token')
+        this.authService.logout(token + '').subscribe((res) => {
+          if (res) {
+            this.authService.deleteToken();
+            this.isLogged = this.cookieService.check('token_access');
+          }
+        })
       }
     })
   }
